move hex to own folder
